+++ src/App.js
@@ -2,9 +2,10 @@ import React, {Component} from 'react';
 import './styles/App.css'
 import {connect} from 'react-redux'
 import {getCurrentUser} from './actions/authUser'
+import {getCookieCards} from './actions/cookieCards'
 import Login from './components/Login'
 import Logout from './components/Logout'
-import NavBar from './components/NavBar'
+import NavList from './components/NavList'
 import {
   withRouter
 } from "react-router-dom";
@@ -14,8 +15,10 @@ class App extends Component {
 
 // add check to see if user is logged in
 // load cards in componentDidMount
+// getcards action on mount
   componentDidMount(){
     this.props.getCurrentUser()
+    this.props.getCookieCards()
   }
 
   render() {
@@ -25,7 +28,7 @@ class App extends Component {
         <div className="App">
         {loggedIn ? <Logout /> : null}
         {this.props.user ? <strong> Welcome {this.props.user.name}</strong> : ""}
-          <NavBar cards={this.props.cards}/>
+          <NavList cards={this.props.cards}/>
         </div>
       );
     }
@@ -39,4 +42,4 @@ class App extends Component {
     }
   }
 
-export default connect(mapState, {getCurrentUser})(App)
+export default connect(mapState, {getCurrentUser,getCookieCards})(App)
