+++ src/components/NavList.js
@@ -1,32 +1,55 @@
-import React from 'react'
+import React, {Component} from 'react'
 import {
+  Route,
+  Switch,
   NavLink,
+  withRouter
 } from "react-router-dom";
-import '../styles/NavBar.css'
+import {connect} from 'react-redux'
+import Login from './Login'
+import Signup from './Signup'
+import CookieCardMaker from './CookieCardMaker'
+import CookieCard from './CookieCard'
+import CookieCardListContainer from '../containers/CookieCardListContainer'
+import NavBar from './NavBar'
 
-const NavList = () => {
-  return (
-    <div className="NavBar">
-        <div className="Nav-Routes">
-          <nav>
-            <ul className="Nav-Route-List">
-              <li>
-                <NavLink to="/login">Log In </NavLink>
-              </li>
-              <li>
-                <NavLink to="/signup">Sign Up</NavLink>
-              </li>
-              <li>
-                <NavLink to="/cookiecards"> My Cards </NavLink>
-              </li>
-              <li>
-                <NavLink to="/cookiecards/new"> Create cookie card  </NavLink>
-              </li>
-            </ul>
-          </nav>
-        </div>
-    </div>
-  )
+
+const NavList = ({cards})  => {
+  {
+    return (
+      <div className="NavBar">
+        <NavBar/>
+        <Switch>
+
+          <Route path="/signup" component={Signup}/>
+
+          <Route path="/login" component={Login}/>
+
+          <Route exact path="/cookiecards" component={CookieCardListContainer}/>
+
+          <Route exact path="/cookiecards/new" component={CookieCardMaker}/>
+
+          <Route exact path='/cookiecards/:id' render={(props) =>{
+            console.log("child cards:",cards);
+            console.log("props: ",props);
+            const card = cards.find(card => card.id === parseInt(props.match.params.id))
+            console.log(card);
+            return <div> <CookieCard {...props} recipe_name={card.recipe_name}
+            recipe_ingredients={card.recipe_ingredients}
+            recipe_steps={card.recipe_steps}/></div>
+
+          }}/>
+
+          <Route exact path='/cookiecards/:id/edit' render={(props) =>{
+
+            return <div> <CookieCard {...props}/></div>
+
+          }}/>
+
+        </Switch>
+      </div>
+    );
+  }
 }
-// this should b navbar
-export default NavList;
+// nav list
+export default NavList
